atualizações novas, verificart

+++ src/courses/courses.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, HttpStatus, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpCode, HttpStatus, NotFoundException, Param, Post, Put } from '@nestjs/common';
 import { CoursesService } from './courses.service';
 
 @Controller('courses')
@@ -11,24 +11,44 @@ export class CoursesController {
     }
 
     @Get(':id')
-    findOne(@Param('id') id: number){
+    async findOne(@Param('id') id: number){
+
+        const cusoExiste = await this.coursesService.findOne(+id);
+
+        if(!cusoExiste){
+                    throw new NotFoundException(`Course Id ${id}, not found`);
+        }
         return this.coursesService.findOne(id);
     }
 
     @Post()
     create(@Body() body){
         return this.coursesService.create(body); 
+
     }
 
     @Put(':id')
-    update(@Param('id') id: number, @Body() body){
+    async update(@Param('id') id: number, @Body() body){
+
+        const cusoExiste = await this.coursesService.findOne(+id);
+
+        if(!cusoExiste){
+                    throw new NotFoundException(`Course Id ${id}, not found`);
+        }
         return this.coursesService.update(id, body);
     }
 
     @HttpCode(HttpStatus.NO_CONTENT)
     @Delete(':id')
     remove(@Param('id') id: number){
+
+        const cusoExiste = this.coursesService.findOne(+id);
+
+        if(!cusoExiste){
+                    throw new NotFoundException(`Course Id ${id}, not found`);
+        }
         return this.coursesService.remove(id);
+
     }
 
 }
